# By João Vicente
# Via João Vicente
* 'master' of https://github.com/captainvera/CGProj:
  Materials

+++ src/LightSource.cpp
@@ -42,7 +42,7 @@ Vector3 & LightSource::getPosition()
 	return _position;
 }
 
-void LightSource::setPosition(GLdouble x, GLdouble y, GLdouble z)
+void LightSource::setPosition(GLfloat x, GLfloat y, GLfloat z)
 {
 	_position.set(x, y, z);
 	GLfloat pos[4] = {x, y, z, _w};
@@ -52,14 +52,14 @@ void LightSource::setPosition(GLdouble x, GLdouble y, GLdouble z)
 void LightSource::setPosition(Vector3 & vec)
 {
 	_position.set(vec);
-	GLfloat pos[4] = { _position._x,
-					   _position._y,
-					   _position._z,
+	GLfloat pos[4] = { (GLfloat)_position._x,
+					   (GLfloat)_position._y,
+					   (GLfloat)_position._z,
 					   0.0 };
 	glLightfv(_num, GL_POSITION, pos);
 }
 
-void LightSource::setDirection(GLdouble x, GLdouble y, GLdouble z)
+void LightSource::setDirection(GLfloat x, GLfloat y, GLfloat z)
 {
 	_direction.set(x, y, z);
 	GLfloat dir[4] = {x ,y ,z ,0.0};
@@ -69,14 +69,14 @@ void LightSource::setDirection(GLdouble x, GLdouble y, GLdouble z)
 void LightSource::setDirection(Vector3 & dir)
 {
 	_direction.set(dir);
-	GLfloat vec[4] = { _direction._x,
-						_direction._y, 
-						_direction._z,
-						0.0 };
+	GLfloat vec[4] = { (GLfloat) _direction._x,
+                       (GLfloat) _direction._y,
+                       (GLfloat) _direction._z,
+                       0.0 };
 	glLightfv(_num, GL_SPOT_DIRECTION, vec);
 }
 
-void LightSource::setAmbient(GLdouble r, GLdouble g, GLdouble b, GLdouble a)
+void LightSource::setAmbient(GLfloat r, GLfloat g, GLfloat b, GLfloat a)
 {
 	_ambient[0] = r;
 	_ambient[1] = g;
@@ -85,7 +85,7 @@ void LightSource::setAmbient(GLdouble r, GLdouble g, GLdouble b, GLdouble a)
 	glLightfv(_num, GL_AMBIENT, _ambient);
 }
 
-void LightSource::setDiffuse(GLdouble r, GLdouble g, GLdouble b, GLdouble a)
+void LightSource::setDiffuse(GLfloat r, GLfloat g, GLfloat b, GLfloat a)
 {
 	_diffuse[0] = r;
 	_diffuse[1] = g;
@@ -94,7 +94,7 @@ void LightSource::setDiffuse(GLdouble r, GLdouble g, GLdouble b, GLdouble a)
 	glLightfv(_num, GL_DIFFUSE, _diffuse);
 }
  
-void LightSource::setSpecular(GLdouble r, GLdouble g, GLdouble b, GLdouble a)
+void LightSource::setSpecular(GLfloat r, GLfloat g, GLfloat b, GLfloat a)
 {
 	_specular[0] = r;
 	_specular[1] = g;
@@ -103,21 +103,21 @@ void LightSource::setSpecular(GLdouble r, GLdouble g, GLdouble b, GLdouble a)
 	glLightfv(_num, GL_SPECULAR, _specular);
 }
 
-void LightSource::setExponent(GLdouble exp)
+void LightSource::setExponent(GLfloat exp)
 {
 	_exponent = exp;
 	GLfloat val[1] = { exp };
 	glLightfv(_num, GL_SPOT_EXPONENT, val);
 }
 
-void LightSource::setCutoff(GLdouble cutoff)
+void LightSource::setCutoff(GLfloat cutoff)
 {
 	_cut_off = cutoff;
 	GLfloat val[1] = { cutoff };
 	glLightfv(_num, GL_SPOT_CUTOFF, val);
 }
 
-void LightSource::setAttenuation(GLdouble constant, GLdouble linear, GLdouble quadric)
+void LightSource::setAttenuation(GLfloat constant, GLfloat linear, GLfloat quadric)
 {
 	GLfloat val[1] = { constant };
 	glLightfv(_num, GL_CONSTANT_ATTENUATION, val);
